+++ CHANGELOG.md
@@ -9,11 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Capture network logs - [#121](https://github.com/ripe-tech/ripe-rainbow/issues/121)
+*
 
 ### Changed
 
 *
+
+### Fixed
+
+*
+
+## [0.10.0] - 2022-02-01
+
+### Added
+
+* Capture network logs - [#121](https://github.com/ripe-tech/ripe-rainbow/issues/121)
 
 ### Fixed